+++ R/downloading_tools.R
@@ -13,7 +13,7 @@
 #' }
 
 download_nhdplusv2 <- function(outdir,
-                               url = paste0("https://s3.amazonaws.com/nhdplus/NHDPlusV21/",
+                               url = paste0("https://s3.amazonaws.com/edap-nhdplus/NHDPlusV21/",
                                             "Data/NationalData/NHDPlusV21_NationalData_Seamless",
                                             "_Geodatabase_Lower48_07.7z")) {